CP-6697: Add "nosubdir" flag for NFSSR
The normal NFSSR creates a sub directory under the NFS directory
given by params "server:serverpath", and use that sub directory
as the SR's working directory.

This patch set add an extra flag "sm-config:nosubdir", which
defaults to false. But when "nosubdir=true", NFSSR will avoid
creating that sub directory. Instead the SR will mount and use
the "serverpath" directory as the working directly.

The purpose of this patch is to provide a behavior compatible
with CloudStack's current NFSSR fork (which always use the
"nosubdir" semantics), so that they can move to the official
NFSSR with the help of this flag.

The reasons for choosing sm-config instead of other key/value
maps are:

  * device-config is RO map, so one can not change it's value
without migrating all VMs away in order to unplug the PBD
and recreate one . This will add considerable burden to CS's
already quite complicated upgrade process from the NFSSR fork
to the new one.

  * other-config is RW map, but it's not a param for sr-create.
So setting nosubdir during sr-create is no-op. One must then
reset the flag and unplug/plug the PBD to make it effective.

This patch also takes care of some corner case when someone
creates both a standard NFSSR and a "nosubdir" NFSSR under the
same serverpath. Ideally, they shouldn't interfere with each
other. This could be useful if CloudStack wants to gradually
moave all their current nosubdir NFSSRs to standard NFSSRs for
their primary storage.

The patch is partially based on a previous version implemented by
Anthony Xu <Xuefei.Xu@citrix.com>.

PS. the fix for an issue (CA-125137) caused by a previous version
of this patch is now integrated in this version.

+++ drivers/NFSSR.py
@@ -75,6 +75,15 @@ class NFSSR(FileSR.FileSR):
         if not self.dconf.has_key('server'):
             raise xs_errors.XenError('ConfigServerMissing')
         self.remoteserver = self.dconf['server']
+        self.nosubdir = False
+        if self.sr_ref and self.session is not None :
+            sm_conf = self.session.xenapi.SR.get_sm_config(self.sr_ref)
+        else:
+            sm_conf = self.srcmd.params.get('sr_sm_config') or {}
+        self.nosubdir = sm_conf.get('nosubdir') == "true"
+        if self.dconf.has_key('serverpath'):
+            self.remotepath = os.path.join(self.dconf['serverpath'],
+                                           not self.nosubdir and sr_uuid or "")
         self.path = os.path.join(SR.MOUNT_BASE, sr_uuid)
 
         # Test for the optional 'nfsoptions' dconf attribute
@@ -113,7 +122,6 @@ class NFSSR(FileSR.FileSR):
 
     def attach(self, sr_uuid):
         self.validate_remotepath(False)
-        self.remotepath = os.path.join(self.dconf['serverpath'], sr_uuid)
         util._testHost(self.dconf['server'], NFSPORT, 'NFSTarget')
         io_timeout = util.get_nfs_timeout(self.session, sr_uuid)
         self.mount_remotepath(sr_uuid, io_timeout)
@@ -136,7 +144,7 @@ class NFSSR(FileSR.FileSR):
 
         temppath = os.path.join(SR.MOUNT_BASE, PROBE_MOUNTPOINT)
 
-        self.mount(temppath, self.dconf['serverpath'])
+        self.mount(temppath, self.remotepath)
         try:
             return nfs.scan_srlist(temppath)
         finally:
@@ -182,20 +190,21 @@ class NFSSR(FileSR.FileSR):
                 pass
             raise exn
 
-        newpath = os.path.join(self.path, sr_uuid)
-        if util.ioretry(lambda: util.pathexists(newpath)):
-            if len(util.ioretry(lambda: util.listdir(newpath))) != 0:
-                self.detach(sr_uuid)
-                raise xs_errors.XenError('SRExists')
-        else:
-            try:
-                util.ioretry(lambda: util.makedirs(newpath))
-            except util.CommandException, inst:
-                if inst.code != errno.EEXIST:
+        if not self.nosubdir:
+            newpath = os.path.join(self.path, sr_uuid)
+            if util.ioretry(lambda: util.pathexists(newpath)):
+                if len(util.ioretry(lambda: util.listdir(newpath))) != 0:
                     self.detach(sr_uuid)
-                    raise xs_errors.XenError('NFSCreate', 
-                        opterr='remote directory creation error is %d' 
-                        % inst.code)
+                    raise xs_errors.XenError('SRExists')
+            else:
+                try:
+                    util.ioretry(lambda: util.makedirs(newpath))
+                except util.CommandException, inst:
+                    if inst.code != errno.EEXIST:
+                        self.detach(sr_uuid)
+                        raise xs_errors.XenError('NFSCreate',
+                            opterr='remote directory creation error is %d'
+                            % inst.code)
         self.detach(sr_uuid)
 
     def delete(self, sr_uuid):
@@ -209,10 +218,10 @@ class NFSSR(FileSR.FileSR):
             # so that we can remove the target SR directory
             self.remotepath = self.dconf['serverpath']
             self.mount_remotepath(sr_uuid)
-            newpath = os.path.join(self.path, sr_uuid)
-
-            if util.ioretry(lambda: util.pathexists(newpath)):
-                util.ioretry(lambda: os.rmdir(newpath))
+            if not self.nosubdir:
+                newpath = os.path.join(self.path, sr_uuid)
+                if util.ioretry(lambda: util.pathexists(newpath)):
+                    util.ioretry(lambda: os.rmdir(newpath))
             self.detach(sr_uuid)
         except util.CommandException, inst:
             self.detach(sr_uuid)
